Moved widgets to components
