+++ src/components/widgets/POS.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import { defineComponent } from "@vue/runtime-core";
-import Base from "@/widgets/Base.vue";
+import { Base } from "./";
 
 export default defineComponent({
 	methods: {
