+++ src/components/widgets/index.ts
@@ -1,0 +1,4 @@
+export { default as Base } from "./Base.vue";
+export { default as Pauze } from "./Pauze.vue";
+export { default as POS } from "./POS.vue";
+export { default as TwitchFollow } from "./TwitchFollow.vue";
